extract module colorpalette middleware
