+++ client/src/index.js
@@ -5,11 +5,9 @@ import { Provider } from "react-redux"
 import App from "./App.js"
 import { rootReducer } from "./reducers.js"
 import { toolMiddleware } from "./middlewares.js"
+import { colorPaletteMiddleware } from "./colorpalette-middleware.js"
 import {
     SET_DATASETS,
-    SET_PALETTE_NAME,
-    SET_PALETTE_NUMBER,
-    SET_PALETTES,
     NEXT_TIME_INDEX,
     PREVIOUS_TIME_INDEX
 } from "./action-types.js"
@@ -17,12 +15,8 @@ import {
     set_url,
     set_dataset,
     set_datasets,
-    set_palette,
-    set_palettes,
     set_palette_name,
-    set_palette_names,
     set_palette_number,
-    set_palette_numbers,
     set_playing,
     set_limits,
     set_time_index,
@@ -70,50 +64,6 @@ let animationMiddleware = store => next => action => {
     }
 }
 
-let colorPaletteMiddleware = store => next => action => {
-    if (action.type == SET_PALETTE_NAME) {
-        // Async get palette numbers
-        let name = action.payload
-        let state = store.getState()
-        if (typeof state.palettes !== "undefined") {
-            let numbers = state.palettes
-                .filter((p) => p.name == name)
-                .map((p) => parseInt(p.number))
-                .concat()
-                .sort((a, b) => a - b)
-            let action = set_palette_numbers(numbers)
-            store.dispatch(action)
-        }
-    }
-    else if (action.type == SET_PALETTE_NUMBER) {
-        // Async get palette numbers
-        let state = store.getState()
-        let name = state.palette_name
-        let number = action.payload
-        if (typeof state.palettes !== "undefined") {
-            let palettes = getPalettes(state.palettes, name, number)
-            if (palettes.length > 0) {
-                let action = set_palette(palettes[0].palette)
-                store.dispatch(action)
-            }
-        }
-    }
-    else if (action.type == SET_PALETTES) {
-        // Set initial palette to Blues 256
-        next(action)
-        next(set_palette_name("Blues"))
-        next(set_palette_number(256))
-        let palettes = getPalettes(action.payload, "Blues", 256)
-        if (palettes.length > 0) {
-            let action = set_palette(palettes[0].palette)
-            next(action)
-        }
-        return
-    }
-
-    return next(action)
-}
-
 let datasetsMiddleware = store => next => action => {
     next(action)
     if (action.type == SET_DATASETS) {
@@ -130,13 +80,6 @@ let mod = function(a, n) {
     // Builtin % operator allows negatives, e.g. -2 % 5 -> -2
     return ((a % n) + n) % n
 }
-
-let getPalettes = function(palettes, name, number) {
-    return palettes
-        .filter((p) => p.name === name)
-        .filter((p) => parseInt(p.number) === parseInt(number))
-}
-
 
 
 window.main = function(baseURL) {
